test_requirements.txt
-e .
PyContracts
caniusepython3
celery
coverage
django-crum
edx-lint
mock
pylint
pytz
sortedcontainers
tox
xblock
responses
